minor changes: formating READ.me, edit types of argarse, fix legend of animation

+++ runsim.py
@@ -21,13 +21,13 @@ def main(*args):
 
     """
     parser = argparse.ArgumentParser(description='Modelling a pandemic with matrices')
-    parser.add_argument('--infection', metavar='P', type=int, default=0.1,
+    parser.add_argument('--infection', metavar='P', type=float, default=0.1,
                         help='this the chance a person infects an adjacent person in a day')
     parser.add_argument('--duration', metavar='N', type=int, default=14,
                         help='days until removed')
     parser.add_argument('--death', metavar='P', type=float, default=0.2,
                         help='the chance an infected person dies')
-    parser.add_argument('--edge', metavar='N', type=float, default=40,
+    parser.add_argument('--edge', metavar='N', type=int, default=40,
                         help='size of side of square matrix')
     parser.add_argument('--file_1', metavar='N', type=str, default=None,
                         help='Filename to be saved of the SIRD graph of the last day')
